Use log2, log2f from C++ std library. These functions are guaranteed to be part of C++11 standard library.
* configure.ac: Remove AC_CHECK_FUNCS for log2, log2f.

* lo-mappers.cc: Delegate to std::log2 or std::log2f for log2.

+++ liboctave/numeric/lo-mappers.cc
@@ -222,7 +222,7 @@ namespace octave
     }
 
     double log2 (double x) { return std::log2 (x); }
-    float log2 (float x) { return std::log2 (x); }
+    float log2 (float x) { return std::log2f (x); }
 
     Complex
     log2 (const Complex& x)
